ai-agent/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
langchain>=0.1.0
langgraph>=0.0.40
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
appwrite>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
pdfplumber>=0.10.0
numpy>=1.24.0
langsmith>=0.1.0
